successfuly seed db
