+++ app/models/spread_model.py
@@ -16,10 +16,10 @@ class Spread(db.Model):
     layout = db.relationship('SpreadLayout')
     position_meanings = db.Column(db.JSON, nullable=False)
 
-    def __init__(self, name, number_of_cards, layout_id, position_meanings):
+    def __init__(self, name, number_of_cards, layout, position_meanings):
         self.name = name
         self.number_of_cards = number_of_cards
-        self.layout_id = layout_id
+        self.layout = layout
         self.position_meanings = position_meanings
 
     def __repr__(self):
